some fixes; especially 64 bit related problems

+++ trunk/src/mstl/m_algobase.h
@@ -33,7 +33,7 @@ template <typename ForwardIterator, typename T>
 void fill(ForwardIterator first, ForwardIterator last, T const& value);
 
 template <typename OutputIterator, typename T>
-OutputIterator fill_n(OutputIterator first, size_t count, T const& value);
+OutputIterator fill_n(OutputIterator first, unsigned count, T const& value);
 
 template <typename BidirectionalIterator>
 void reverse(BidirectionalIterator first, BidirectionalIterator last);
